+++ src/front/component/routing/Router.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import {Routes, Route, BrowserRouter } from "react-router-dom"
-import { Landing } from '../Landing'
+import { Landing } from '../../Landing'
+
 
 
 export const Router = () => {
