+++ src/front/pages/Footer.js
